Extract core algorithm to a library
